requirements.txt
torch==2.1.0
faiss-cpu==1.7.4
nbformat==5.9.2
python-dotenv==1.0.0
GitPython==3.1.40
llama-index==0.9.48
sentence-transformers==2.2.2
openai==1.3.5
anthropic==0.8.1
google-generativeai==0.3.2
